Add imx8qm Display VM configuration
Added a configuration file to build the imx8qm image with the
requirements to launch the Display VM with GPU/HDMI passthrough.
This configuration file adds the imx8qm_display_vm.nix overlay file
which will have all the required overlays.

In this commit the overlay to support the GPU/HDMI passthrough from
u-boot was also added.

Signed-off-by: Juan Pablo Ruiz <juanpablo.ruiz@tii.ae>